refactor(structure): Move the files into further folders for better project structure.
